pyproject.toml
[project]
name = "colouranga"
version = "0.1.0"
description = "Contains code of the master's thesis on the subject of 'Diffusion models implementation to improve color uniformity of manga characters'."
requires-python = "==3.10.*"
readme = "README.md"
authors = [{ name = "eovsianik", email = "liza.ovsianik@gmail.com" }]
license = { text = "MIT" }
dependencies = [
    "diffusers==0.21.*",
    "accelerate",
    "stable_diffusion_reference_only @ git+https://github.com/aihao2000/stable-diffusion-reference-only.git",
    "anime-segmentation @ git+https://github.com/aihao2000/anime-segmentation.git",
    "torch==2.3.*",
    "opencv-python>=4.10.0.84",
    "torchmetrics>=1.4.1",
    "numpy>=2.1.1",
    "pillow>=10.4.0",
    "torchinfo>=1.8.0",
    "transformers>=4.44.2",
    "datasets>=2.21.0",
    "huggingface-hub>=0.24.6",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true


[[tool.pdm.source]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
verify_ssl = true
include_packages = ["torch", "torchvision"]



[tool.pdm.dev-dependencies]
jupyter = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
]
[tool.ruff]
src = ["libs"]
line-length = 100
target-version = "py311"
include = ["*.py", "*.pyi", "**/pyproject.toml", "*.ipynb"]

[tool.ruff.lint]
per-file-ignores = { "*.ipynb" = [
    "F401",
    "E402",
    "PLE1142",
    "PLR2004",
    "PLR0913",
], "*.pyi" = [
    "F401",
] }
select = [
    # "D",   # pydocstyle
    # "ANN", # flake8-annotations
    "TCH", # flake8-type-checking
    "FBT", # flake8-boolean-trap

    "E",    # pycodestyle
    "F",    # pyflakes
    "UP",   # pyupgrade
    "I",    # isort
    "PL",   # pylint
    "RUF",  # ruff-specific
    "N",    # pep8-naming
    "C4",   # flake8-comprehensions
    "B",    # flake8-bugbear
    "PTH",  # flake8-use-pathlib
    "TRY",  # tryceratops
    "PERF", # Perflint 
    "FURB", # refurb 
    "PT",   # flake8-pytest-style
]
ignore = [
    "RUF001", # ambiguous cyrillic letters
    "RUF002", # ambiguous cyrillic letters
    "RUF003", # ambiguous cyrillic letters
    "ANN101", # missing-type-self #? deprecated
    "N812",
    "D202",   # no-blank-line-after-function
]

[tool.pyright]
pythonVersion = "3.10"
reportPrivateImportUsage = false
reportMissingImports = true
# reportMissingTypeStubs = true


#? cSpell Configuration:
# cspell: words tryceratops FURB ipywidgets
